Restructuring to make a build and install possible.
