+++ setup.py
@@ -123,7 +123,8 @@ setup(
     #
     #   py_modules=["my_module"],
     #
-    packages=find_packages(exclude=["contrib", "docs", "tests"]),  # Required
+    packages=find_packages(where='src', exclude=["contrib", "docs", "tests"]),  # Required
+    package_dir={"": "src"},
     # Specify which Python versions you support. In contrast to the
     # 'Programming Language' classifiers above, 'pip install' will check this
     # and refuse to install the project if the version does not match. If you
